Network.Haskoin.Node

Description

This package provides basic types used for the Bitcoin networking protocol together with Data.Binary instances for efficiently serializing and de-serializing them. More information on the bitcoin protocol is available here: http://en.bitcoin.it/wiki/Protocol_specification

Requesting data

data GetData

The GetData type is used to retrieve information on a specific object (Block or Tx) identified by the objects hash. The payload of a GetData request is a list of InvVector which represent all the hashes for which a node wants to request information. The response to a GetBlock message wille be either a Block or a Tx message depending on the type of the object referenced by the hash. Usually, GetData messages are sent after a node receives an Inv message to obtain information on unknown object hashes.

Constructors

GetData

Fields

getDataList :: ![InvVector]

List of object hashes

data Inv

Inv messages are used by nodes to advertise their knowledge of new objects by publishing a list of hashes. Inv messages can be sent unsolicited or in response to a GetBlocks message.

Constructors

Inv

Fields

invList :: ![InvVector]

Inventory vectors

data InvVector

Invectory vectors represent hashes identifying objects such as a Block or a Tx. They are sent inside messages to notify other peers about new data or data they have requested.

Constructors

InvVector

Fields

invType :: !InvType

Type of the object referenced by this inventory vector

invHash :: !Hash256

Hash of the object referenced by this inventory vector

data InvType

Data type identifying the type of an inventory vector.

Constructors

InvError
Error. Data containing this type can be ignored.

InvTx
InvVector hash is related to a transaction

InvBlock
InvVector hash is related to a block

InvMerkleBlock
InvVector has is related to a merkle block

data NotFound

A NotFound message is returned as a response to a GetData message whe one of the requested objects could not be retrieved. This could happen, for example, if a tranasaction was requested and was not available in the memory pool of the receiving node.

Constructors

NotFound

Fields

notFoundList :: ![InvVector]

Inventory vectors related to this request

Network types

newtype VarInt

Data type representing a variable length integer. The VarInt type usually precedes an array or a string that can vary in length.

Constructors

VarInt

Fields

getVarInt :: Word64

newtype VarString

Data type for variable length strings. Variable length strings are serialized as a VarInt followed by a bytestring.

Constructors

VarString

Fields

getVarString :: ByteString

data NetworkAddress

Data type describing a bitcoin network address. Addresses are stored in IPv6. IPv4 addresses are mapped to IPv6 using IPv4 mapped IPv6 addresses: http://en.wikipedia.org/wiki/IPv6#IPv4-mapped_IPv6_addresses. Sometimes, timestamps are sent together with the NetworkAddress such as in the Addr data type.

Constructors

NetworkAddress

Fields

naServices :: !Word64

Bitmask of services available for this address

naAddress :: !SockAddr

IPv6 address and port

data Addr

Provides information on known nodes in the bitcoin network. An Addr type is sent inside a Message as a response to a GetAddr message.

Constructors

Addr

Fields

addrList :: ![NetworkAddressTime]

type NetworkAddressTime = (Word32, NetworkAddress)

Network address with a timestamp

data Version

When a bitcoin node creates an outgoing connection to another node, the first message it will send is a Version message. The other node will similarly respond with it's own Version message.

Constructors

Version

Fields

version :: !Word32

Protocol version being used by the node.

services :: !Word64

Bitmask of features to enable for this connection.

timestamp :: !Word64

UNIX timestamp

addrRecv :: !NetworkAddress

Network address of the node receiving this message.

addrSend :: !NetworkAddress

Network address of the node sending this message.

verNonce :: !Word64

Randomly generated identifying sent with every version message. This nonce is used to detect connection to self.

userAgent :: !VarString

User agent

startHeight :: !Word32

The height of the last block received by the sending node.

relay :: !Bool

Wether the remote peer should announce relaying transactions or not. This feature is enabled since version >= 70001. See BIP37 for more details.

newtype Ping

A Ping message is sent to bitcoin peers to check if a TCP/IP connection is still valid.

Constructors

Ping

Fields

pingNonce :: Word64

A random nonce used to identify the recipient of the ping request once a Pong response is received.

newtype Pong

A Pong message is sent as a response to a ping message.

Constructors

Pong

Fields

pongNonce :: Word64

When responding to a Ping request, the nonce from the Ping is copied in the Pong response.

data Alert

Data type describing signed messages that can be sent between bitcoin nodes to display important notifications to end users about the health of the network.

Constructors

Alert

Fields

alertPayload :: !VarString

Alert payload.

alertSignature :: !VarString

ECDSA signature of the payload

data Reject

The reject message is sent when messages are rejected by a peer.

Constructors

Reject

Fields

rejectMessage :: !MessageCommand

Type of message rejected

rejectCode :: !RejectCode

Code related to the rejected message

rejectReason :: !VarString

Text version of rejected reason

rejectData :: !ByteString

Optional extra data provided by some errors

data RejectCode

Constructors

RejectMalformed

RejectInvalid

RejectObsolete

RejectDuplicate

RejectNonStandard

RejectDust

RejectInsufficientFee

RejectCheckpoint

reject :: MessageCommand -> RejectCode -> ByteString -> Reject

Convenience function to build a Reject message

Messages

data Message

The Message type is used to identify all the valid messages that can be sent between bitcoin peers. Only values of type Message will be accepted by other bitcoin peers as bitcoin protocol messages need to be correctly serialized with message headers. Serializing a Message value will include the MessageHeader with the correct checksum value automatically. No need to add the MessageHeader separately.

Constructors

MVersion !Version

MVerAck

MAddr !Addr

MInv !Inv

MGetData !GetData

MNotFound !NotFound

MGetBlocks !GetBlocks

MGetHeaders !GetHeaders

MTx !Tx

MBlock !Block

MMerkleBlock !MerkleBlock

MHeaders !Headers

MGetAddr

MFilterLoad !FilterLoad

MFilterAdd !FilterAdd

MFilterClear

MPing !Ping

MPong !Pong

MAlert !Alert

MMempool

MReject !Reject

data MessageHeader

Data type representing the header of a Message. All messages sent between nodes contain a message header.

Constructors

MessageHeader

Fields

headMagic :: !Word32

Network magic bytes. It is used to differentiate messages meant for different bitcoin networks, such as prodnet and testnet.

headCmd :: !MessageCommand

Message command identifying the type of message. included in the payload.

headPayloadSize :: !Word32

Byte length of the payload.

headChecksum :: !CheckSum32

Checksum of the payload.

data MessageCommand

A MessageCommand is included in a MessageHeader in order to identify the type of message present in the payload. This allows the message de-serialization code to know how to decode a particular message payload. Every valid Message constructor has a corresponding MessageCommand constructor.

Constructors

MCVersion

MCVerAck

MCAddr

MCInv

MCGetData

MCNotFound

MCGetBlocks

MCGetHeaders

MCTx

MCBlock

MCMerkleBlock

MCHeaders

MCGetAddr

MCFilterLoad

MCFilterAdd

MCFilterClear

MCPing

MCPong

MCAlert

MCMempool

MCReject

Bloom filters

data BloomFlags

The bloom flags are used to tell the remote peer how to auto-update the provided bloom filter.

Constructors

BloomUpdateNone
Never update

BloomUpdateAll
Auto-update on all outputs

BloomUpdateP2PubKeyOnly
Only auto-update on outputs that are pay-to-pubkey or pay-to-multisig. This is the default setting.

data BloomFilter

A bloom filter is a probabilistic data structure that SPV clients send to other peers to filter the set of transactions received from them. Bloom filters are probabilistic and have a false positive rate. Some transactions that pass the filter may not be relevant to the receiving peer. By controlling the false positive rate, SPV nodes can trade off bandwidth versus privacy.

Constructors

BloomFilter

Fields

bloomData :: !(Seq Word8)

Bloom filter data

bloomHashFuncs :: !Word32

Number of hash functions for this filter

bloomTweak :: !Word32

Hash function random nonce

bloomFlags :: !BloomFlags

Bloom filter auto-update flags

newtype FilterLoad

Set a new bloom filter on the peer connection.

Constructors

FilterLoad

Fields

filterLoadBloomFilter :: BloomFilter

newtype FilterAdd

Add the given data element to the connections current filter without requiring a completely new one to be set.

Constructors

FilterAdd

Fields

getFilterData :: ByteString

bloomCreate

Arguments

:: Int
Number of elements

-> Double
False positive rate

-> Word32
A random nonce (tweak) for the hash function. It should be a random number but the secureness of the random value is not of geat consequence.

-> BloomFlags
Bloom filter flags

-> BloomFilter
Bloom filter

Build a bloom filter that will provide the given false positive rate when the given number of elements have been inserted.

bloomInsert

Arguments

:: BloomFilter
Original bloom filter

-> ByteString
New data to insert

-> BloomFilter
Bloom filter containing the new data

Insert arbitrary data into a bloom filter. Returns the new bloom filter containing the new data.

bloomContains

Arguments

:: BloomFilter
Bloom filter

-> ByteString
Data that will be checked against the given bloom filter

-> Bool
Returns True if the data matches the filter

Tests if some arbitrary data matches the filter. This can be either because the data was inserted into the filter or because it is a false positive.

isBloomValid

Arguments

:: BloomFilter
Bloom filter to test

-> Bool
True if the given filter is valid

Tests if a given bloom filter is valid.

isBloomEmpty :: BloomFilter -> Bool

Returns True if the filter is empty (all bytes set to 0x00)

isBloomFull :: BloomFilter -> Bool

Returns True if the filter is full (all bytes set to 0xff)