Network.Haskoin.Test

Description

This package provides test types for Network.Haskoin

Util Arbitrary instances

data ArbitraryByteString

Arbitrary strict ByteString

Constructors

ArbitraryByteString ByteString

data ArbitraryNotNullByteString

Arbitrary strict ByteString that is not empty

Constructors

ArbitraryNotNullByteString ByteString

newtype ArbitraryUTCTime

Arbitrary UTCTime that generates dates after 01 Jan 1970 01:00:00 CET

Constructors

ArbitraryUTCTime UTCTime

newtype ArbitraryHash512

Constructors

ArbitraryHash512 Hash512

newtype ArbitraryHash256

Constructors

ArbitraryHash256 Hash256

newtype ArbitraryHash160

Constructors

ArbitraryHash160 Hash160

newtype ArbitraryCheckSum32

Constructors

ArbitraryCheckSum32 CheckSum32

Crypto Arbitrary instances

newtype ArbitraryPrvKey

Arbitrary private key (can be both compressed or uncompressed)

Constructors

ArbitraryPrvKey PrvKey

newtype ArbitraryPrvKeyC

Arbitrary compressed private key

Constructors

ArbitraryPrvKeyC PrvKeyC

newtype ArbitraryPrvKeyU

Arbitrary uncompressed private key

Constructors

ArbitraryPrvKeyU PrvKeyU

data ArbitraryPubKey

Arbitrary public key (can be both compressed or uncompressed) with its corresponding private key.

Constructors

ArbitraryPubKey PrvKey PubKey

data ArbitraryPubKeyC

Arbitrary compressed public key with its corresponding private key.

Constructors

ArbitraryPubKeyC PrvKeyC PubKeyC

data ArbitraryPubKeyU

Arbitrary uncompressed public key with its corresponding private key.

Constructors

ArbitraryPubKeyU PrvKeyU PubKeyU

newtype ArbitraryAddress

Arbitrary address (can be a pubkey or script hash address)

Constructors

ArbitraryAddress Address

newtype ArbitraryPubKeyAddress

Arbitrary public key hash address

Constructors

ArbitraryPubKeyAddress Address

newtype ArbitraryScriptAddress

Arbitrary script hash address

Constructors

ArbitraryScriptAddress Address

data ArbitrarySignature

Arbitrary message hash, private key, nonce and corresponding signature. The signature is generated with a random message, random private key and a random nonce.

Constructors

ArbitrarySignature Hash256 PrvKey Signature

data ArbitraryXPrvKey

Arbitrary extended private key.

Constructors

ArbitraryXPrvKey XPrvKey

data ArbitraryXPubKey

Arbitrary extended public key with its corresponding private key.

Constructors

ArbitraryXPubKey XPrvKey XPubKey

data ArbitraryHardPath

Constructors

ArbitraryHardPath HardPath

data ArbitrarySoftPath

Constructors

ArbitrarySoftPath SoftPath

data ArbitraryDerivPath

Constructors

ArbitraryDerivPath DerivPath

data ArbitraryParsedPath

Constructors

ArbitraryParsedPath ParsedPath

Node Arbitrary instances

newtype ArbitraryVarInt

Arbitrary VarInt

Constructors

ArbitraryVarInt VarInt

newtype ArbitraryVarString

Arbitrary VarString

Constructors

ArbitraryVarString VarString

newtype ArbitraryNetworkAddress

Arbitrary NetworkAddress

Constructors

ArbitraryNetworkAddress NetworkAddress

newtype ArbitraryNetworkAddressTime

Arbitrary NetworkAddressTime

Constructors

ArbitraryNetworkAddressTime (Word32, NetworkAddress)

newtype ArbitraryInvType

Arbitrary InvType

Constructors

ArbitraryInvType InvType

newtype ArbitraryInvVector

Arbitrary InvVector

Constructors

ArbitraryInvVector InvVector

newtype ArbitraryInv

Arbitrary non-empty Inv

Constructors

ArbitraryInv Inv

newtype ArbitraryVersion

Arbitrary Version

Constructors

ArbitraryVersion Version

newtype ArbitraryAddr

Arbitrary non-empty Addr

Constructors

ArbitraryAddr Addr

newtype ArbitraryAlert

Arbitrary alert with random payload and signature. Signature is not valid.

Constructors

ArbitraryAlert Alert

newtype ArbitraryReject

Arbitrary Reject

Constructors

ArbitraryReject Reject

newtype ArbitraryRejectCode

Arbitrary RejectCode

Constructors

ArbitraryRejectCode RejectCode

newtype ArbitraryGetData

Arbitrary non-empty GetData

Constructors

ArbitraryGetData GetData

newtype ArbitraryNotFound

Arbitrary NotFound

Constructors

ArbitraryNotFound NotFound

newtype ArbitraryPing

Arbitrary Ping

Constructors

ArbitraryPing Ping

newtype ArbitraryPong

Arbitrary Pong

Constructors

ArbitraryPong Pong

data ArbitraryBloomFlags

Arbitrary bloom filter flags

Constructors

ArbitraryBloomFlags BloomFlags

data ArbitraryBloomFilter

Arbitrary bloom filter with its corresponding number of elements and false positive rate.

Constructors

ArbitraryBloomFilter Int Double BloomFilter

data ArbitraryFilterLoad

Arbitrary FilterLoad

Constructors

ArbitraryFilterLoad FilterLoad

data ArbitraryFilterAdd

Arbitrary FilterAdd

Constructors

ArbitraryFilterAdd FilterAdd

newtype ArbitraryMessageCommand

Arbitrary MessageCommand

Constructors

ArbitraryMessageCommand MessageCommand

Message Arbitrary instances

newtype ArbitraryMessageHeader

Arbitrary MessageHeader

Constructors

ArbitraryMessageHeader MessageHeader

newtype ArbitraryMessage

Arbitrary Message

Constructors

ArbitraryMessage Message

Script Arbitrary instances

newtype ArbitraryScriptOp

Arbitrary ScriptOp (push operations have random data)

Constructors

ArbitraryScriptOp ScriptOp

newtype ArbitraryScript

Arbitrary Script with random script ops

Constructors

ArbitraryScript Script

newtype ArbitraryIntScriptOp

Arbtirary ScriptOp with a value in [OP_1 .. OP_16]

Constructors

ArbitraryIntScriptOp ScriptOp

newtype ArbitraryPushDataType

Arbitrary PushDataType

Constructors

ArbitraryPushDataType PushDataType

data ArbitraryTxSignature

Arbitrary message hash, private key and corresponding TxSignature. The signature is generated deterministically using a random message and a random private key.

Constructors

ArbitraryTxSignature TxHash PrvKey TxSignature

newtype ArbitrarySigHash

Arbitrary SigHash (including invalid/unknown sighash codes)

Constructors

ArbitrarySigHash SigHash

newtype ArbitraryValidSigHash

Arbitrary valid SigHash

Constructors

ArbitraryValidSigHash SigHash

data ArbitraryMSParam

Arbitrary m of n parameters

Constructors

ArbitraryMSParam Int Int

newtype ArbitraryScriptOutput

Arbitrary ScriptOutput (Can by any valid type)

Constructors

ArbitraryScriptOutput ScriptOutput

newtype ArbitrarySimpleOutput

Arbitrary ScriptOutput of type PayPK, PayPKHash or PayMS (Not PayScriptHash)

Constructors

ArbitrarySimpleOutput ScriptOutput

newtype ArbitraryPKOutput

Arbitrary ScriptOutput of type PayPK

Constructors

ArbitraryPKOutput ScriptOutput

newtype ArbitraryPKHashOutput

Arbitrary ScriptOutput of type PayPKHash

Constructors

ArbitraryPKHashOutput ScriptOutput

newtype ArbitraryMSOutput

Arbitrary ScriptOutput of type PayMS

Constructors

ArbitraryMSOutput ScriptOutput

newtype ArbitraryMSCOutput

Arbitrary ScriptOutput of type PayMS containing only compressed keys

Constructors

ArbitraryMSCOutput ScriptOutput

newtype ArbitrarySHOutput

Arbitrary ScriptOutput of type PayScriptHash

Constructors

ArbitrarySHOutput ScriptOutput

newtype ArbitraryScriptInput

Arbitrary ScriptInput

Constructors

ArbitraryScriptInput ScriptInput

newtype ArbitrarySimpleInput

Arbitrary ScriptInput of type SpendPK, SpendPKHash or SpendMulSig (not ScriptHashInput)

Constructors

ArbitrarySimpleInput ScriptInput

newtype ArbitraryPKInput

Arbitrary ScriptInput of type SpendPK

Constructors

ArbitraryPKInput ScriptInput

newtype ArbitraryPKHashInput

Arbitrary ScriptInput of type SpendPK

Constructors

ArbitraryPKHashInput ScriptInput

newtype ArbitraryPKHashCInput

Arbitrary ScriptInput of type SpendPK with a compressed public key

Constructors

ArbitraryPKHashCInput ScriptInput

newtype ArbitraryMSInput

Arbitrary ScriptInput of type SpendMulSig

Constructors

ArbitraryMSInput ScriptInput

newtype ArbitrarySHInput

Arbitrary ScriptInput of type ScriptHashInput

Constructors

ArbitrarySHInput ScriptInput

newtype ArbitraryMulSigSHCInput

Arbitrary ScriptInput of type ScriptHashInput containing a RedeemScript of type PayMulSig and an input of type SpendMulSig. Only compressed keys are used.

Constructors

ArbitraryMulSigSHCInput ScriptInput

Transaction Arbitrary instances

newtype ArbitrarySatoshi

Arbitrary amount of Satoshi as Word64 (Between 1 and 21e14)

Constructors

ArbitrarySatoshi Word64

newtype ArbitraryTx

Arbitrary Tx

Constructors

ArbitraryTx Tx

newtype ArbitraryTxHash

Constructors

ArbitraryTxHash TxHash

newtype ArbitraryTxIn

Arbitrary TxIn

Constructors

ArbitraryTxIn TxIn

newtype ArbitraryTxOut

Arbitrary TxOut

Constructors

ArbitraryTxOut TxOut

newtype ArbitraryOutPoint

Arbitrary OutPoint

Constructors

ArbitraryOutPoint OutPoint

newtype ArbitraryCoinbaseTx

Arbitrary CoinbaseTx

Constructors

ArbitraryCoinbaseTx CoinbaseTx

newtype ArbitraryAddrOnlyTx

Arbitrary Tx containing only inputs of type SpendPKHash, SpendScriptHash (multisig) and outputs of type PayPKHash and PaySH. Only compressed public keys are used.

Constructors

ArbitraryAddrOnlyTx Tx

newtype ArbitraryAddrOnlyTxIn

Arbitrary TxIn that can only be of type SpendPKHash or SpendScriptHash (multisig). Only compressed public keys are used.

Constructors

ArbitraryAddrOnlyTxIn TxIn

newtype ArbitraryAddrOnlyTxOut

Arbitrary TxOut that can only be of type PayPKHash or PaySH

Constructors

ArbitraryAddrOnlyTxOut TxOut

data ArbitrarySigInput

Arbitrary SigInput with the corresponding private keys used to generate the ScriptOutput or RedeemScript

Constructors

ArbitrarySigInput SigInput [PrvKey]

data ArbitraryPKSigInput

Arbitrary SigInput with a ScriptOutput of type PayPK

Constructors

ArbitraryPKSigInput SigInput PrvKey

data ArbitraryPKHashSigInput

Arbitrary SigInput with a ScriptOutput of type PayPKHash

Constructors

ArbitraryPKHashSigInput SigInput PrvKey

data ArbitraryMSSigInput

Arbitrary SigInput with a ScriptOutput of type PayMulSig

Constructors

ArbitraryMSSigInput SigInput [PrvKey]

data ArbitrarySHSigInput

Arbitrary SigInput with ScriptOutput of type PaySH and a RedeemScript

Constructors

ArbitrarySHSigInput SigInput [PrvKey]

data ArbitrarySigningData

Arbitrary Tx (empty TxIn), SigInputs and PrvKeys that can be passed to signTx or detSignTx to fully sign the Tx.

Constructors

ArbitrarySigningData Tx [SigInput] [PrvKey]

data ArbitraryPartialTxs

Constructors

ArbitraryPartialTxs [Tx] [(ScriptOutput, OutPoint, Int, Int)]

Block Arbitrary instances

newtype ArbitraryBlock

Arbitrary Block

Constructors

ArbitraryBlock Block

newtype ArbitraryBlockHeader

Arbitrary BlockHeader

Constructors

ArbitraryBlockHeader BlockHeader

newtype ArbitraryBlockHash

Constructors

ArbitraryBlockHash BlockHash

newtype ArbitraryGetBlocks

Arbitrary GetBlocks

Constructors

ArbitraryGetBlocks GetBlocks

newtype ArbitraryGetHeaders

Arbitrary GetHeaders

Constructors

ArbitraryGetHeaders GetHeaders

newtype ArbitraryHeaders

Arbitrary Headers

Constructors

ArbitraryHeaders Headers

newtype ArbitraryMerkleBlock

Arbitrary MerkleBlock

Constructors

ArbitraryMerkleBlock MerkleBlock